@@ -3,81 +3,86 @@
   "version": "0.0.1",
   "repository": {
     "type": "git",
-    "url": "https://github.com/photoview/photoview.git",
+    "url": "https://github.com/kkovaletp/photoview.git",
     "directory": "ui"
   },
   "license": "GPL-3.0",
   "description": "UI app for Photoview",
   "scripts": {
-    "start": "vite --host=0.0.0.0",
+    "start": "vite",
     "mon": "nodemon",
     "build": "vite build",
     "lint": "eslint ./src --max-warnings 0 --config .eslintrc.js",
+    "lint:ci": "eslint ./src --config .eslintrc.js -f json > eslint-report.json",
     "test": "vitest",
-    "test:ci": "CI=true vitest --reporter verbose --run --coverage",
+    "test:ci": "CI=true vitest --reporter=junit --reporter=verbose --run --coverage",
     "genSchemaTypes": "apollo client:codegen --target=typescript --globalTypesFile=src/__generated__/globalTypes.ts --passthroughCustomScalars && prettier --write */**/__generated__/*.ts",
     "extractTranslations": "i18next -c i18next-parser.config.js",
     "prepare": "(cd .. && ./ui/node_modules/.bin/husky install)"
   },
   "dependencies": {
-    "@apollo/client": "^3.6.9",
-    "@babel/preset-typescript": "^7.18.6",
-    "@headlessui/react": "^1.6.6",
-    "@types/geojson": "^7946.0.8",
-    "@types/jest": "^28.1.4",
-    "@types/mapbox-gl": "^2.7.3",
-    "@types/react": "^18.0.15",
-    "@types/react-dom": "^18.0.6",
-    "@types/react-helmet": "^6.1.5",
+    "@apollo/client": "~3.6.10",
+    "@babel/preset-typescript": "^7.24.7",
+    "@headlessui/react": "^2.1.2",
+    "@mapbox/mapbox-gl-language": "^1.0.0",
+    "@types/geojson": "^7946.0.14",
+    "@types/jest": "^29.5.12",
+    "@types/mapbox-gl": "^3.4.0",
+    "@types/react": "^18.3.3",
+    "@types/react-dom": "^18.3.0",
+    "@types/react-helmet": "^6.1.11",
     "@types/react-router-dom": "^5.3.3",
-    "@types/styled-components": "^5.1.25",
-    "@types/url-join": "^4.0.1",
+    "@types/styled-components": "^5.1.34",
+    "@types/url-join": "^4.0.3",
     "@vitejs/plugin-react": "^1.3.2",
-    "autoprefixer": "^10.4.7",
+    "autoprefixer": "^10.4.20",
     "blurhash": "^1.1.5",
-    "classnames": "^2.3.1",
+    "classnames": "^2.5.1",
     "connect-history-api-fallback": "^2.0.0",
-    "copy-to-clipboard": "^3.3.1",
-    "i18next": "^21.8.13",
-    "mapbox-gl": "^2.9.1",
-    "postcss": "^8.4.39",
-    "prettier": "^2.7.1",
-    "react": "^18.2.0",
+    "copy-to-clipboard": "^3.3.3",
+    "i18next": "^21.10.0",
+    "mapbox-gl": "~3.4.0",
+    "postcss": "^8.4.41",
+    "prettier": "^2.8.8",
+    "react": "^18.3.1",
     "react-blurhash": "^0.1.3",
-    "react-dom": "^18.2.0",
+    "react-dom": "^18.3.1",
     "react-helmet": "^6.1.0",
-    "react-hook-form": "^7.33.1",
-    "react-i18next": "^11.18.0",
+    "react-hook-form": "^7.52.2",
+    "react-i18next": "^12.3.1",
     "react-router-dom": "^6.3.0",
     "react-swipeable": "^7.0.0",
-    "react-test-renderer": "^18.2.0",
-    "styled-components": "^5.3.5",
+    "react-test-renderer": "^18.3.1",
+    "styled-components": "^5.3.11",
     "subscriptions-transport-ws": "^0.11.0",
     "tailwind-override": "^0.6.1",
     "tailwindcss": "npm:@tailwindcss/postcss7-compat@^2.2.17",
     "url-join": "^5.0.0",
-    "vite": "^2.9.13",
-    "vite-plugin-svgr": "^2.2.0"
+    "vite": "^2.9.18",
+    "vite-plugin-svgr": "^2.4.0"
   },
   "devDependencies": {
-    "@testing-library/jest-dom": "^5.16.4",
-    "@testing-library/react": "^13.3.0",
-    "@testing-library/user-event": "^14.2.1",
-    "@typescript-eslint/eslint-plugin": "^5.30.6",
-    "@typescript-eslint/parser": "^5.30.6",
-    "@vitest/ui": "^0.17.1",
+    "@testing-library/jest-dom": "^6.4.8",
+    "@testing-library/react": "^15.0.7",
+    "@testing-library/user-event": "^14.5.2",
+    "@typescript-eslint/eslint-plugin": "^8.0.1",
+    "@typescript-eslint/parser": "^8.0.1",
+    "@vitest/coverage-c8": "^0.33.0",
+    "@vitest/ui": "^0.34.7",
     "apollo": "2.34.0",
     "apollo-language-server": "1.26.9",
-    "c8": "^7.11.3",
-    "eslint": "^8.19.0",
-    "eslint-config-prettier": "^8.5.0",
-    "husky": "^8.0.1",
-    "i18next-parser": "^8.13.0",
-    "lint-staged": "^13.0.3",
+    "c8": "^10.1.2",
+    "eslint": "^8.57.0",
+    "eslint-config-prettier": "^9.1.0",
+    "eslint-plugin-react": "^7.35.0",
+    "eslint-plugin-react-hooks": "^4.6.2",
+    "husky": "^9.1.4",
+    "i18next-parser": "^9.0.2",
+    "lint-staged": "^15.2.8",
     "nodemon": "^3.1.4",
     "ts-node": "^10.9.2",
-    "typescript": "^5.5.3",
-    "vitest": "^0.17.1"
+    "typescript": "^5.5.4",
+    "vitest": "^0.30.1"
   },
   "overrides": {
     "graphql": "^15.5.0"
